runtime: Don't call regular method if assignment hash matches
This is a pretty nasty breakage in lookup code that I somehow didn't
catch in tests, so I'm adding one for it now. This caused all unary
methods that had a single-keyword variant to ignore the single-keyword
variant and only use the unary one. This also broke argument sentinel
checks but I never noticed it. Ouch!