runtime: Simplify slot writing code
MapBuilder handles a lot of edge cases that are guaranteed to not happen
when we're creating objects ex-nihilo. In particular:
- Slot names are guaranteed to be unique
- Argument slots always come last

We cache some values on ObjectDescriptor and create the object with a
single pass. This nets us a nice 4% boost in the Fibonacci example.

MapBuilder will probably go away after this because it only exists for a
single use-case now (_AddSlots:).