runtime: Refactor Interpreter to use labelled switch continue
This greatly simplifies the interpreter code since we no longer have to
rely on tail call semantics. Labelled switch continue was added to Zig
for this exact purpose, so let's use it.

In doing so, I also reduced our reliance on context.zig (I dislike it at
this point and want to get rid of it) and turned Interpreter into an
actual struct with methods (instead of passing a ton of arguments
around).

Just simplifying things gives us a cool 7-8% performance boost on the
Fibonacci example.