runtime: Remodel stack sentinels
Previously we compared against a magic value for the stack sentinel,
which sucked because:
- It required us to shoehorn a magic value into each type.
- It didn't allow us to put any interesting information into
  sentinels.

The new model puts the sentinels out-of-band, and only enables them as
values when explicitly asked for, which should be a little more
efficient. When disabled, all relevant values become voids meaning they
take no memory, and all sentinel-related operations are compiled out.

We additionally now store a unique value for each sentinel, verifying
that the sentinel we popped is actually matching the one we pushed.