runtime: Implement GetLocal/PutLocal instructions
This is the first step towards improving the performance of the VM
greatly.

One of the major sources of slowdowns for us was local slots in
activations requiring full lookups. Even though at parse time we know
exactly which self-sends would hit slots in the activation, we would
still emit `Send`/`SelfSend` messages, which would run the lookup
machinery. Now we simply get the relevant assignable slot value from the
current activation object which ends up executing way less code.

There is still a lot of low-hanging fruit. The target here is to remove
the activation object entirely and rely on the Zig side for activation
handling instead.

This gives us a huge performance boost, with 18.5% less wall clock time
on the Fibonacci example.