runtime: Introduce the locals stack + use it for (inline) methods/blocks
This is a bit of a large commit because the changes are all
interconnected. The main idea is that now, instead of getting locals
from the activation object, we have a "locals stack" that is responsible
for storing all of the locals that are used by the various activations.

Every time a method is activated, the method needs to allocate some
space on the locals stack. This space is the total count of all locals
used by the method itself, and any inline methods/blocks defined
within (as blocks/inline methods also have access to the method's
locals).

To calculate all of the locals required, AstGen now has the
ObjectDescriptorLink struct which is weaved through code generation and
calculates the local index for each local that's accessible from a given
activation within a containing method. The sum is then recorded on the
CreateMethod instruction, and when the method is activated, the locals
stack gets the memory reserved.

The GetLocal/PutLocal instructions now offset into the locals stack
instead of using the activation object. In essence, we no longer use any
of the assignable slot values of the activation object itself anymore.

This gives us -15% wall clock time on the Fibonacci example, on top of
the -18.5% wall clock time with the initial locals implementation,
bringing the total performance boost of locals to -31% or 1.45x which is
pretty huge.

Note that because we still create the activation object, we currently
have to copy the assignable and argument slot values to the locals
stack, so there is still a lot of performance gains to be had. When
constants are obtainable without going through the activation object,
and the activation object is completely removed, the arguments will be
used directly off the argument stack instead of being copied which
should bring a further performance boost.