runtime: Make executables method-level + add CreateInlineMethod insn.
The purpose of this change is to prepare for further optimizations that
can come from making things more local.

As part of this change, I also took the liberty of properly splitting
the creation and execution of inline methods from non-inline ones since
their execution now diverges in non-trivial ways.

No measurable performance change in the Fibonacci example.